feat(layout): refactor AdminLayout to use SideBar and AdminHeader components; rename AvatarDemo to UserAvatar
