+++ src/components/admin/FooterBar.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 import Pagination from "@/components/Pagination";
 import { ChevronDown } from "lucide-react";
